+++ include/H2DE/objects/H2DE_object.h
@@ -2,6 +2,7 @@
 #define H2DE_OBJECT_H
 
 #include <H2DE/engine/H2DE_engine.h>
+#include <H2DE/engine/H2DE_error.h>
 class H2DE_Engine;
 class H2DE_Texture;
 class H2DE_Sprite;
@@ -42,7 +43,20 @@ protected:
     }
     void removeSurface(std::unordered_map<std::string, H2DE_Surface*>& surfaces, const std::string& name);
 
-    static H2DE_Surface* getSurface(const std::unordered_map<std::string, H2DE_Surface*>& surfaces, const std::string& name);
+    template<typename H2DE_Surface_T>
+    static H2DE_Surface_T* getSurface(const std::unordered_map<std::string, H2DE_Surface*>& surfaces, const std::string& name) {
+        auto it = surfaces.find(name);
+        if (it == surfaces.end()) {
+            H2DE_Error::throwError("Surface named \"" + name + "\" not found");
+        }
+
+        H2DE_Surface_T* castedSurface = dynamic_cast<H2DE_Surface_T*>(it->second);
+        if (castedSurface == nullptr) {
+            H2DE_Error::throwError("Can't cast surface \"" + name +  "\" to template type");
+        }
+
+        return castedSurface;
+    }
 
     static const std::vector<H2DE_Surface*> getSortedSurfaces(std::unordered_map<std::string, H2DE_Surface*>& surfaces);
     static const std::array<H2DE_Translate, 4> getCorners(const H2DE_Transform& transform);